Move FreeThreadedHen to separate files
The AtlHen files became a bit big.
