+++ AtlServer/AtlHen.h
@@ -25,37 +25,6 @@ private:
 
 OBJECT_ENTRY_AUTO(CLSID_AtlHen, AtlHen)
 
-
-
-class ATL_NO_VTABLE FreeThreadedHen :
-    public CComObjectRootEx<CComMultiThreadModel>,
-    public CComCoClass<FreeThreadedHen, &CLSID_FreeThreadedHen>,
-    public IHen
-{
-public:
-    DECLARE_REGISTRY_RESOURCEID(IDR_HEN)
-    DECLARE_PROTECT_FINAL_CONSTRUCT()
-    DECLARE_GET_CONTROLLING_UNKNOWN()
-
-    BEGIN_COM_MAP(FreeThreadedHen)
-        COM_INTERFACE_ENTRY(IHen)
-        COM_INTERFACE_ENTRY_AGGREGATE(IID_IMarshal, m_marshaler)
-    END_COM_MAP()
-
-    HRESULT FinalConstruct();
-
-    HRESULT Cluck() override;
-    HRESULT CluckAsync(IAsyncCluckObserver* cluckObserver) override;
-
-private:
-    CComPtr<IUnknown> m_marshaler;
-    unsigned long m_myThreadId = GetCurrentThreadId();
-};
-
-OBJECT_ENTRY_AUTO(CLSID_FreeThreadedHen, FreeThreadedHen)
-
-
-
 class ATL_NO_VTABLE AtlCluckObserver : public CComObjectRootEx<CComSingleThreadModel>,
     public CComCoClass<AtlCluckObserver, &CLSID_AtlCluckObserver>,
     public IAsyncCluckObserver
