+++ AtlServer/AtlServer.vcxproj
@@ -21,6 +21,7 @@
   <ItemGroup>
     <ClCompile Include="AtlHen.cpp" />
     <ClCompile Include="AtlServer.cpp" />
+    <ClCompile Include="FreeThreadedHen.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
@@ -35,6 +36,7 @@
   <ItemGroup>
     <ClInclude Include="AtlHen.h" />
     <ClInclude Include="framework.h" />
+    <ClInclude Include="FreeThreadedHen.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="targetver.h" />
